terra-website-clone
Description: --This project is a clone of the landing page of the Terra Learning Management System (LMS) site, created as part of a weekly assignment. The goal was to replicate the design and key elements of the Terra LMS landing page, providing a visually appealing and functional introduction to the platform.
Co-Authored-By: udy1111 <169306742+udy1111@users.noreply.github.com>
Co-Authored-By: bickyblinkz30 <58190649+bickyblinkz30@users.noreply.github.com>